[#139][#138] Removes the read more link but keeps the content
This commit makes the 'read more' empty text so it does not display
or create an empty paragraph around it.
By keeping the content rather than the excerpt we retain HTML in the
display, so can keep e.g. links to presenters
To govern the length of the text we can use the <!--more--> break in the
page body.