[#135] Addresses the issue where programmes don't link from schedule
We're trying to create a link from the title of a show to a potential
worppress slug.
I've used the same funtion in Wordpress that creates the slug from
a string.
Hopefiully this should cath most cases where the Title contains a
character that can't be used in the URL.