[#148] Updates how show descriptions are displayed
The Next On formatting was different to any other display of show
data. So pulled that into the show_description function so that
it is more uniform.
Removed the 140 character limit, and just show whatb is there because
many shows have very long descriptions.
Fixed it so that presenter links now show in all next, upcoming, past
listings as well